rag/pdf-rag/requirements.txt
google-genai
numpy
python-dotenv
pypdf